requirements.txt
streamlit
pandas
fastapi
uvicorn
python-dotenv
requests
tqdm
